ctt: Add writers class
The writers are classes supposed to output a rendered job.

The exact endpoint will depend on the writer class being used, but so far
we have the option to write to a file or send it to a LAVA instance.

More could be used for example to dump in on stdout.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>