writers: Handle the case when boards are retired
When boards are retired such as the armada-385-db-ap,
the ci_launcher explode because the board is not available.

Handle retired boards in the same way of "offline": check the
status, print an error and raise an exception.

Note: This error will be handled for every jobs so it could be nice
to improve it and check before sending every job the status of the
board. Thanks to that, jobs will be created and attempted to be sent
only if the board is available.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>