src/writers: return the list of job IDs instead of strings
In preparation for making ctt.py optionally synchronous (i.e wait for
the jobs to complete), this commit changes the LavaWriter()::write()
method to run the list of job IDs instead of the list of URLs to the
jobs in the LAVA interface.

The code in src/crafters.py is adjusted accordingly to continue
printing the URLs.

TODO: fix the test cases in tests/. Maxime: help, help!

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>