notifications: no limit in number of jobs returned
LAVA was updated to have a default limit of 200 jobs returned. We don't
want that. To bypass it, we need to pass None as the limit (in LAVA the
list of returned jobs is computed like this: jobs[:limit]).

jobs[:None] is the only way in Python to return the array jobs in its
entirety.

XMLRPC isn't allowed by default to send None parameters except if we
explicitely ask it.

Signed-off-by: Quentin Schulz <quentin.schulz@bootlin.com>