notifications: fix nand test always reported as failed
Using several lava-test-set/case instead of one results in test cases
being named after lava-test-case call.

Thus, we cannot look for "nand" in the test cases as nand is actually
the name of the whole test (used only for synchronization purpose in
LAVA; it does not appear in results) and not the name of the test case.

Rather match on the TestSuite being named with the suffix custom-tests
since it's a common identifier for our custom tests.

Signed-off-by: Quentin Schulz <quentin.schulz@bootlin.com>